requirements.txt
discord.py-self==2.0.1
groq==0.25.0
loguru==0.7.3
python-dotenv==1.1.0
redis==6.1.0
hiredis==3.1.1
